// app/bot/interfaces/WorldData.ts
export interface AFG {
    continent: string;
    location: string;
    last_updated_date: string;
    total_cases: number;
    new_cases: number;
    new_cases_smoothed: number;
    total_deaths: number;
    new_deaths: number;
    new_deaths_smoothed: number;
    total_cases_per_million: number;
    new_cases_per_million: number;
    new_cases_smoothed_per_million: number;
    total_deaths_per_million: number;
    new_deaths_per_million: number;
    new_deaths_smoothed_per_million: number;
    reproduction_rate: number;
    icu_patients?: any;
    icu_patients_per_million?: any;
    hosp_patients?: any;
    hosp_patients_per_million?: any;
    weekly_icu_admissions?: any;
    weekly_icu_admissions_per_million?: any;
    weekly_hosp_admissions?: any;
    weekly_hosp_admissions_per_million?: any;
    total_tests?: any;
    new_tests?: any;
    total_tests_per_thousand?: any;
    new_tests_per_thousand?: any;
    new_tests_smoothed: number;
    new_tests_smoothed_per_thousand: number;
    positive_rate: number;
    tests_per_case: number;
    tests_units: string;
    total_vaccinations: number;
    people_vaccinated: number;
    people_fully_vaccinated: number;
    total_boosters?: any;
    new_vaccinations?: any;
    new_vaccinations_smoothed: number;
    total_vaccinations_per_hundred: number;
    people_vaccinated_per_hundred: number;
    people_fully_vaccinated_per_hundred: number;
    total_boosters_per_hundred?: any;
    new_vaccinations_smoothed_per_million: number;
    new_people_vaccinated_smoothed: number;
    new_people_vaccinated_smoothed_per_hundred: number;
    stringency_index: number;
    population: number;
    population_density: number;
    median_age: number;
    aged_65_older: number;
    aged_70_older: number;
    gdp_per_capita: number;
    extreme_poverty?: any;
    cardiovasc_death_rate: number;
    diabetes_prevalence: number;
    female_smokers?: any;
    male_smokers?: any;
    handwashing_facilities: number;
    hospital_beds_per_thousand: number;
    life_expectancy: number;
    human_development_index: number;
    excess_mortality_cumulative_absolute?: any;
    excess_mortality_cumulative?: any;
    excess_mortality?: any;
    excess_mortality_cumulative_per_million?: any;
}

export interface OWIDAFR {
    continent?: any;
    location: string;
    last_updated_date: string;
    total_cases: number;
    new_cases: number;
    new_cases_smoothed: number;
    total_deaths: number;
    new_deaths: number;
    new_deaths_smoothed: number;
    total_cases_per_million: number;
    new_cases_per_million: number;
    new_cases_smoothed_per_million: number;
    total_deaths_per_million: number;
    new_deaths_per_million: number;
    new_deaths_smoothed_per_million: number;
    reproduction_rate?: any;
    icu_patients?: any;
    icu_patients_per_million?: any;
    hosp_patients?: any;
    hosp_patients_per_million?: any;
    weekly_icu_admissions?: any;
    weekly_icu_admissions_per_million?: any;
    weekly_hosp_admissions?: any;
    weekly_hosp_admissions_per_million?: any;
    total_tests?: any;
    new_tests?: any;
    total_tests_per_thousand?: any;
    new_tests_per_thousand?: any;
    new_tests_smoothed?: any;
    new_tests_smoothed_per_thousand?: any;
    positive_rate?: any;
    tests_per_case?: any;
    tests_units?: any;
    total_vaccinations: number;
    people_vaccinated: number;
    people_fully_vaccinated: number;
    total_boosters: number;
    new_vaccinations: number;
    new_vaccinations_smoothed: number;
    total_vaccinations_per_hundred: number;
    people_vaccinated_per_hundred: number;
    people_fully_vaccinated_per_hundred: number;
    total_boosters_per_hundred: number;
    new_vaccinations_smoothed_per_million: number;
    new_people_vaccinated_smoothed: number;
    new_people_vaccinated_smoothed_per_hundred: number;
    stringency_index?: any;
    population: number;
    population_density?: any;
    median_age?: any;
    aged_65_older?: any;
    aged_70_older?: any;
    gdp_per_capita?: any;
    extreme_poverty?: any;
    cardiovasc_death_rate?: any;
    diabetes_prevalence?: any;
    female_smokers?: any;
    male_smokers?: any;
    handwashing_facilities?: any;
    hospital_beds_per_thousand?: any;
    life_expectancy?: any;
    human_development_index?: any;
    excess_mortality_cumulative_absolute?: any;
    excess_mortality_cumulative?: any;
    excess_mortality?: any;
    excess_mortality_cumulative_per_million?: any;
}

export interface ALB {
    continent: string;
    location: string;
    last_updated_date: string;
    total_cases: number;
    new_cases: number;
    new_cases_smoothed: number;
    total_deaths: number;
    new_deaths: number;
    new_deaths_smoothed: number;
    total_cases_per_million: number;
    new_cases_per_million: number;
    new_cases_smoothed_per_million: number;
    total_deaths_per_million: number;
    new_deaths_per_million: number;
    new_deaths_smoothed_per_million: number;
    reproduction_rate: number;
    icu_patients?: any;
    icu_patients_per_million?: any;
    hosp_patients?: any;
    hosp_patients_per_million?: any;
    weekly_icu_admissions?: any;
    weekly_icu_admissions_per_million?: any;
    weekly_hosp_admissions?: any;
    weekly_hosp_admissions_per_million?: any;
    total_tests: number;
    new_tests: number;
    total_tests_per_thousand: number;
    new_tests_per_thousand: number;
    new_tests_smoothed: number;
    new_tests_smoothed_per_thousand: number;
    positive_rate: number;
    tests_per_case: number;
    tests_units: string;
    total_vaccinations?: any;
    people_vaccinated?: any;
    people_fully_vaccinated?: any;
    total_boosters?: any;
    new_vaccinations?: any;
    new_vaccinations_smoothed?: any;
    total_vaccinations_per_hundred?: any;
    people_vaccinated_per_hundred?: any;
    people_fully_vaccinated_per_hundred?: any;
    total_boosters_per_hundred?: any;
    new_vaccinations_smoothed_per_million?: any;
    new_people_vaccinated_smoothed?: any;
    new_people_vaccinated_smoothed_per_hundred?: any;
    stringency_index?: any;
    population: number;
    population_density: number;
    median_age: number;
    aged_65_older: number;
    aged_70_older: number;
    gdp_per_capita: number;
    extreme_poverty: number;
    cardiovasc_death_rate: number;
    diabetes_prevalence: number;
    female_smokers: number;
    male_smokers: number;
    handwashing_facilities?: any;
    hospital_beds_per_thousand: number;
    life_expectancy: number;
    human_development_index: number;
    excess_mortality_cumulative_absolute?: any;
    excess_mortality_cumulative?: any;
    excess_mortality?: any;
    excess_mortality_cumulative_per_million?: any;
}
// ...
export interface WorldData {
    AFG: AFG;
    OWID_AFR: OWIDAFR;
    ALB: ALB;
    DZA: DZA;
    AND: AND;
    AGO: AGO;
    AIA: AIA;
    ATG: ATG;
    ARG: ARG;
    ARM: ARM;
    ABW: ABW;
    OWID_ASI: OWIDASI;
    AUS: AUS;
    AUT: AUT;
    AZE: AZE;
    BHS: BHS;
    BHR: BHR;
    BGD: BGD;
    BRB: BRB;
    BLR: BLR;
    BEL: BEL;
    BLZ: BLZ;
    BEN: BEN;
    BMU: BMU;
    BTN: BTN;
    BOL: BOL;
    BES: BES;
    BIH: BIH;
    BWA: BWA;
    BRA: BRA;
    VGB: VGB;
    BRN: BRN;
    BGR: BGR;
    BFA: BFA;
    BDI: BDI;
    KHM: KHM;
    CMR: CMR;
    CAN: CAN;
    CPV: CPV;
    CYM: CYM;
    CAF: CAF;
    TCD: TCD;
    CHL: CHL;
    CHN: CHN;
    COL: COL;
    COM: COM;
    COG: COG;
    COK: COK;
    CRI: CRI;
    CIV: CIV;
    HRV: HRV;
    CUB: CUB;
    CUW: CUW;
    CYP: CYP;
    CZE: CZE;
    COD: COD;
    DNK: DNK;
    DJI: DJI;
    DMA: DMA;
    DOM: DOM;
    ECU: ECU;
    EGY: EGY;
    SLV: SLV;
    GNQ: GNQ;
    ERI: ERI;
    EST: EST;
    SWZ: SWZ;
    ETH: ETH;
    OWID_EUR: OWIDEUR;
    OWID_EUN: OWIDEUN;
    FRO: FRO;
    FLK: FLK;
    FJI: FJI;
    FIN: FIN;
    FRA: FRA;
    PYF: PYF;
    GAB: GAB;
    GMB: GMB;
    GEO: GEO;
    DEU: DEU;
    GHA: GHA;
    GIB: GIB;
    GRC: GRC;
    GRL: GRL;
    GRD: GRD;
    GUM: GUM;
    GTM: GTM;
    GGY: GGY;
    GIN: GIN;
    GNB: GNB;
    GUY: GUY;
    HTI: HTI;
    OWID_HIC: OWIDHIC;
    HND: HND;
    HKG: HKG;
    HUN: HUN;
    ISL: ISL;
    IND: IND;
    IDN: IDN;
    OWID_INT: OWIDINT;
    IRN: IRN;
    IRQ: IRQ;
    IRL: IRL;
    IMN: IMN;
    ISR: ISR;
    ITA: ITA;
    JAM: JAM;
    JPN: JPN;
    JEY: JEY;
    JOR: JOR;
    KAZ: KAZ;
    KEN: KEN;
    KIR: KIR;
    OWID_KOS: OWIDKOS;
    KWT: KWT;
    KGZ: KGZ;
    LAO: LAO;
    LVA: LVA;
    LBN: LBN;
    LSO: LSO;
    LBR: LBR;
    LBY: LBY;
    LIE: LIE;
    LTU: LTU;
    OWID_LIC: OWIDLIC;
    OWID_LMC: OWIDLMC;
    LUX: LUX;
    MAC: MAC;
    MDG: MDG;
    MWI: MWI;
    MYS: MYS;
    MDV: MDV;
    MLI: MLI;
    MLT: MLT;
    MHL: MHL;
    MRT: MRT;
    MUS: MUS;
    MEX: MEX;
    FSM: FSM;
    MDA: MDA;
    MCO: MCO;
    MNG: MNG;
    MNE: MNE;
    MSR: MSR;
    MAR: MAR;
    MOZ: MOZ;
    MMR: MMR;
    NAM: NAM;
    NPL: NPL;
    NLD: NLD;
    NCL: NCL;
    NZL: NZL;
    NIC: NIC;
    NER: NER;
    NGA: NGA;
    OWID_NAM: OWIDNAM;
    PRK: PRK;
    MKD: MKD;
    NOR: NOR;
    OWID_OCE: OWIDOCE;
    OMN: OMN;
    PAK: PAK;
    PLW: PLW;
    PSE: PSE;
    PAN: PAN;
    PNG: PNG;
    PRY: PRY;
    PER: PER;
    PHL: PHL;
    POL: POL;
    PRT: PRT;
    PRI: PRI;
    QAT: QAT;
    ROU: ROU;
    RUS: RUS;
    RWA: RWA;
    SHN: SHN;
    KNA: KNA;
    LCA: LCA;
    SPM: SPM;
    VCT: VCT;
    WSM: WSM;
    SMR: SMR;
    STP: STP;
    SAU: SAU;
    SEN: SEN;
    SRB: SRB;
    SYC: SYC;
    SLE: SLE;
    SGP: SGP;
    SXM: SXM;
    SVK: SVK;
    SVN: SVN;
    SLB: SLB;
    SOM: SOM;
    ZAF: ZAF;
    OWID_SAM: OWIDSAM;
    KOR: KOR;
    SSD: SSD;
    ESP: ESP;
    LKA: LKA;
    SDN: SDN;
    SUR: SUR;
    SWE: SWE;
    CHE: CHE;
    SYR: SYR;
    TWN: TWN;
    TJK: TJK;
    TZA: TZA;
    THA: THA;
    TLS: TLS;
    TGO: TGO;
    TON: TON;
    TTO: TTO;
    TUN: TUN;
    TUR: TUR;
    TCA: TCA;
    UGA: UGA;
    UKR: UKR;
    ARE: ARE;
    GBR: GBR;
    USA: USA;
    VIR: VIR;
    OWID_UMC: OWIDUMC;
    URY: URY;
    UZB: UZB;
    VUT: VUT;
    VAT: VAT;
    VEN: VEN;
    VNM: VNM;
    WLF: WLF;
    OWID_WRL: OWIDWRL;
    YEM: YEM;
    ZMB: ZMB;
    ZWE: ZWE;
}